TestWeb/webApi/requirements.txt
asn1crypto==0.24.0
Babel==2.6.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==1.11.18
django-bootstrap3==11.0.0
django-chunked-upload==1.1.3
django-extra-views==0.11.0
django-filter==2.1.0
django-haystack==2.8.1
django-oscar==1.6.7
django-phonenumber-field==2.0.1
django-ratelimit==1.1.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.4
flake8==3.7.7
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
openapi-codec==1.3.2
pbr==5.1.3
phonenumbers==8.10.9
Pillow==5.3.0
purl==1.5
pycodestyle==2.5.0
pycountry==18.12.8
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
yapf==0.27.0
